Make each list item in friends list an individual component
